Working base turret with neo550s

+++ northstar-robomaster-project/taproot/src/tap/motor/sparkmax/rev_motor_encoder.hpp
@@ -62,7 +62,7 @@ private:
 
     tap::arch::MilliTimeout encoderDisconnectTimeout;
 
-    int16_t shaftRPM;
+    float shaftRPM;
 };
 
 }  // namespace tap::motor
